Continued cleanup
