+++ rt_gene/src/rt_gene/gaze_tools.py
@@ -69,7 +69,7 @@ def get_endpoint(theta, phi, center_x, center_y, length=300):
     return endpoint_x, endpoint_y
 
 
-def get_head_pose(rot_head):
+def limit_yaw(rot_head):
     euler_angles_head = list(tf.transformations.euler_from_quaternion(rot_head))
     # [0]: pos - roll right, neg -   roll left
     # [1]: pos - look down,  neg -   look up
@@ -79,6 +79,73 @@ def get_head_pose(rot_head):
         euler_angles_head[2] -= 2 * np.pi
 
     return euler_angles_head
+
+
+def crop_face_from_image(color_img, box):
+    _bb = list(map(int, box))
+    return color_img[_bb[1]: _bb[3], _bb[0]: _bb[2]]
+
+
+def is_rotation_vector_stable(last_rotation_vector, current_rotation_vector):
+    # check to see if rotation_vector is wild, if so, stop checking head positions
+    _unit_rotation_vector = current_rotation_vector / np.linalg.norm(current_rotation_vector)
+    _unit_last_rotation_vector = last_rotation_vector / np.linalg.norm(last_rotation_vector)
+    _theta = np.arccos(np.dot(_unit_last_rotation_vector.reshape(3, ), _unit_rotation_vector))
+    # tqdm.write("Head Rotation from last frame: {:.2f}".format(_theta))
+    if _theta > 0.1:
+        # we have too much rotation here, likely unstable, thus error out
+        print('Could not estimate head pose due to instability of landmarks')
+        return False
+    else:
+        return True
+
+
+def move_box(box, offset):
+    """Move the box to direction specified by vector offset"""
+    left_x = box[0] + offset[0]
+    top_y = box[1] + offset[1]
+    right_x = box[2] + offset[0]
+    bottom_y = box[3] + offset[1]
+
+    return [left_x, top_y, right_x, bottom_y]
+
+
+def box_in_image(box, image):
+    """Check if the box is in image"""
+    rows = image.shape[0]
+    cols = image.shape[1]
+
+    return box[0] >= 0 and box[1] >= 0 and box[2] <= cols and box[3] <= rows
+
+
+def get_square_box(box):
+    """Get a square box out of the given box, by expanding it."""
+    left_x = box[0]
+    top_y = box[1]
+    right_x = box[2]
+    bottom_y = box[3]
+
+    box_width = right_x - left_x
+    box_height = bottom_y - top_y
+
+    # Check if box is already a square. If not, make it a square.
+    diff = box_height - box_width
+    delta = int(abs(diff) / 2)
+
+    if diff == 0:  # Already a square.
+        return box
+    elif diff > 0:  # Height > width, a slim box.
+        left_x -= delta
+        right_x += delta
+        if diff % 2 == 1:
+            right_x += 1
+    else:  # Width > height, a short box.
+        top_y -= delta
+        bottom_y += delta
+        if diff % 2 == 1:
+            bottom_y += 1
+
+    return [left_x, top_y, right_x, bottom_y]
 
 
 def angle_loss(y_true, y_pred):
